uploading photo is consolidated on backend
