--- server/api/middleware/parseForm.js
@@ -1,14 +1,0 @@
-import formidable from 'formidable';
-
-//helper function for processing picture through form
-function parseForm(req){
-    return new Promise((resolve, reject) => {
-        const form = formidable({multiples: false});
-        form.parse(req, (err, fields, files) => {
-            if(err) return reject(err);
-            resolve ({fields, files})
-        })
-    });
-}
-
-export { parseForm }
