+++ server/package.json
@@ -16,7 +16,7 @@
     "mongodb": "^6.17.0",
     "node-mailjet": "^6.0.9",
     "passport-local": "^1.0.0",
-    "sharp": "^0.34.3"
+    "sharp": "^0.34.4"
   },
   "devDependencies": {
     "@types/connect-mongo": "^3.1.2",
